+++ packages/abstract-extractor/lib/ergo/network/ExplorerNetwork.ts
@@ -3,6 +3,7 @@ import ergoExplorerClientFactory from '@rosen-clients/ergo-explorer';
 import { BlockInfo } from '../../interfaces';
 import { ErgoBox } from '../interfaces';
 import { AbstractNetwork } from './AbstractNetwork';
+import { V1 } from '@rosen-clients/ergo-explorer';
 
 export class ExplorerNetwork extends AbstractNetwork {
   private api;
@@ -25,6 +26,31 @@ export class ExplorerNetwork extends AbstractNetwork {
   };
 
   /**
+   * convert explorer api boxes to ErgoBox interface
+   * @param box
+   * @returns ErgoBox
+   */
+  convertBox = async (box: V1.OutputInfo): Promise<ErgoBox> => {
+    const spendInfo = box.spentTransactionId
+      ? await this.getTxBlock(box.spentTransactionId)
+      : undefined;
+    return {
+      blockId: box.blockId,
+      boxId: box.boxId,
+      creationHeight: box.creationHeight,
+      inclusionHeight: box.settlementHeight,
+      ergoTree: box.ergoTree,
+      index: box.index,
+      transactionId: box.transactionId,
+      value: box.value,
+      additionalRegisters: box.additionalRegisters,
+      assets: box.assets,
+      spentHeight: spendInfo?.height,
+      spentBlockId: spendInfo?.hash,
+    };
+  };
+
+  /**
    * use explorer api to return related boxes by specified address
    * @param address
    * @param offset
@@ -42,7 +68,11 @@ export class ExplorerNetwork extends AbstractNetwork {
     });
     if (!boxes.items)
       throw new Error('Explorer BoxesByAddress api expected to have items');
-    return { boxes: boxes.items, hasNextBatch: boxes.total > offset + limit };
+    const resultBoxes: Array<ErgoBox> = [];
+    for (const box of boxes.items) {
+      resultBoxes.push(await this.convertBox(box));
+    }
+    return { boxes: resultBoxes, hasNextBatch: boxes.total > offset + limit };
   };
 
   /**
@@ -63,6 +93,10 @@ export class ExplorerNetwork extends AbstractNetwork {
     });
     if (!boxes.items)
       throw new Error('Explorer BoxesByTokeId api expected to have items');
-    return { boxes: boxes.items, hasNextBatch: boxes.total > offset + limit };
+    const resultBoxes: Array<ErgoBox> = [];
+    for (const box of boxes.items) {
+      resultBoxes.push(await this.convertBox(box));
+    }
+    return { boxes: resultBoxes, hasNextBatch: boxes.total > offset + limit };
   };
 }