+++ packages/abstract-extractor/lib/ergo/network/NodeNetwork.ts
@@ -15,21 +15,30 @@ export class NodeNetwork extends AbstractNetwork {
   }
 
   /**
-   * covert node api boxes to ErgoBox interface
+   * convert node api boxes to ErgoBox interface
    * @param box
    * @returns ErgoBox
    */
-  convertToErgoBox = async (box: IndexedErgoBox): Promise<ErgoBox> => ({
-    transactionId: box.transactionId || '',
-    index: box.index || 0,
-    value: box.value || 0n,
-    ergoTree: box.ergoTree || '',
-    creationHeight: box.creationHeight || 0,
-    assets: box.assets || [],
-    additionalRegisters: box.additionalRegisters,
-    boxId: box.boxId || '',
-    blockId: (await this.getTxBlock(box.transactionId!)).hash,
-  });
+  convertToErgoBox = async (box: IndexedErgoBox): Promise<ErgoBox> => {
+    const tx = await this.getTxBlock(box.transactionId!);
+    const spendInfo = box.spentTransactionId
+      ? await this.getTxBlock(box.spentTransactionId)
+      : undefined;
+    return {
+      transactionId: box.transactionId || '',
+      index: box.index || 0,
+      value: box.value || 0n,
+      ergoTree: box.ergoTree || '',
+      creationHeight: box.creationHeight || 0,
+      inclusionHeight: tx.height,
+      assets: box.assets || [],
+      additionalRegisters: box.additionalRegisters,
+      boxId: box.boxId || '',
+      blockId: tx.hash,
+      spentBlockId: spendInfo?.hash ?? undefined,
+      spentHeight: spendInfo?.height ?? undefined,
+    };
+  };
 
   /**
    * return block information of specified tx