+++ packages/abstract-extractor/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @rosen-bridge/abstract-extractor
 
+## 0.1.2
+
+### Patch Changes
+
+- fix spend info extraction in initialize process
+
 ## 0.1.1
 
 ### Patch Changes