+++ packages/evm-address-tx-extractor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rosen-bridge/evm-address-tx-extractor
 
+## 0.1.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@0.1.2
+  - @rosen-bridge/scanner@4.0.2
+
 ## 0.1.4
 
 ### Patch Changes