+++ packages/watcher-data-extractor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @rosen-bridge/watcher-data-extractor
 
+## 6.0.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @rosen-bridge/abstract-extractor@0.1.2
+  - @rosen-bridge/scanner@4.0.2
+
 ## 6.0.1
 
 ### Patch Changes